"Remove WeatherForecast and update queuing system in Post.Query API"
The WeatherForecast API endpoints and related classes have been deleted from the Post.Cmd.Api and Post.Query.Api as it's not longer useful for our application. The queuing system in Post.Query.Api has been updated, by changing the reference from CQRS.Core.Messaging to CQRS.Core.Infrastructure. This is due to the refactoring and renaming of the classes and interfaces from the old messaging module to a new centralized infrastructure module to improve the application architecture and increase the flexibility for future extension and management. This commit also provides the setup for the RabbitMq with hardcoded config and dequeues a message from RabbitMq in the startup of Post.Query.Api. As well, the project reference in SM-post solution file is updated accordingly.