Refactor consumers in infrastructure and removed TTL properties
The application's infrastructure consumers have been refactored to improve the modularity and readability of the code. Message queue consumers were segregated into individual classes by the event type they handle, improving readability while allowing for better separation of concerns. This replaced the existing 'ConsumerHostedService' and 'PostCreatedEventConsumer' which were less modulated and intuitive. The interface 'IQueueMessage' and class 'Message' had the property 'TimeToLive' removed as it was identified to be unnecessary leading to unnecessary complexity in the queue management. The 'PostCreatedEventConsumer' key function was moved into the newly added consumers - 'PostCreatedQueueConsumer', 'MessageUpdatedQueueConsumer', 'PostLikedQueueConsumer', 'CommentRemovedQueueConsumer', 'CommentAddedQueueConsumer', 'CommentUpdatedQueueConsumer', 'PostRemovedQueueConsumer'. A new extensions class 'ConsumerExtensions' was added, and 'AddQueueConsumers' method attached to it for much improved and intuitive consumer management. Queuing logic in the 'QueueProducer' class was also simplified with the removal of TTL properties.  Finally, updates to the program start up configuration in Program.cs reflect the use of these new consumers.