Implement CQRS query side for Post microservice
This commit provides the implementation of the query side (Read side) of the Command Query Responsibility Segregation (CQRS) pattern for the Post microservice. Added the BaseQuery abstract class to provide a basis for custom query classes. Created a PostLookupController to handle all post retrieval HTTP requests. Developed different query classes for retrieving posts based on different conditions (all, by ID, by author, with comments, with likes). In addition, the necessary service dependencies were registered in Program.cs and a QueryDispatcher was added for dispatching the queries to their corresponding handlers. Made a slight adjustment in the PostRepository by correcting the method name ListWIthCommentsAsync to ListWithCommentsAsync. This change will enable efficient data retrieval operations for the Post microservice in a CQRS context.