Add OpenTelemetry for tracing and monitoring
Several references to OpenTelemetry packages have been added to enable tracing and monitoring on Post.Cmd and Post.Query. A new OpenTelemetryConfig class was created, and its values are being configured in the appsettings.json files. Additionally, a Jaeger service was added to the docker-compose file to facilitate visualizing traces.