Update target framework and dependencies versions; Refactor OpenTelemetry setup
The major change in this commit is the update of the target framework of various projects from .NET 6.0 to .NET 8.0, as reflected in the project files. Correspondingly, the versions of multiple package references have also been updated to their compatible versions with .NET 8.0. Additionally, the setup for OpenTelemetry has been refactored with a new static method 'AddOpenTelemetryWithJaeger' in a new 'OpenTelemetryExtensions' class for cleaner and more reusable trace instrumentation setup across projects. Configuration-related changes have also been made in 'appsettings.json' to support these updates.